[Xamarin.Android.Build.Tasks] fix @(LibraryProjectZip) and NuGets
Context: https://github.com/xamarin/GooglePlayServicesComponents/issues/652

In .NET 6, `@(AndroidLibrary)` defaults to `%(Pack)` being true, while
`@(LibraryProjectZip)` mistakenly does not. This would be an issue if
you migrated a project using `@(LibraryProjectZip)` and were not
relying on the default wildcards that would use:

    <AndroidLibrary Include="**/*.aar" />

Let's set this value by default, to ease migration to .NET 6.

I updated a test for this scenario and fixed some assertions that are
working since 2ca2a103 was merged.