[Mono.Android] Copy Pdb and Xml files to the Packs directory
Commit c21e78ca added support for running the unit tests without
the need to run `make pack-dotnet` first. This greatly reduces the
time it takes to test changes locally.

However we are missing some of the `.pdb` files for assemblies like
`Mono.Android` and `Java.Interop`. This causes some test
failures locally because these tests specifically check that these
files are fast deployed during the debug process.

So lets update the relevant code to copy over all the files for
`Mono.Android` and `Java.Interop` rather than just the assemblies.