[MSBuildDeviceIntegration] Make Debugging Test work under .NET 6+. (#7156)
When running the `DebuggingTest.ApplicationRunsWithDebuggerAndBreaks()`
unit test under .NEt 6, we see the following test failure.

	Error Message:
	     Should have hit 2 breakpoints. Only hit 1
	  Expected: 2
	  But was:  1

This is because we are using code like this:

	session.Breakpoints = new BreakpointStore {
	   { Path.Combine (Root, appBuilder.ProjectDirectory, "MainActivity.cs"),  20 },
	}

The line number is hard coded.  Under .NET 6 we seem to be getting a
slightly different file compared to running under Classic: the `using`
statements are missing.  As a result the line numbers have changed.
It turns out that line 20 in the .NET 6 `MainActivity.cs` file is a
blank line, not the `base.OnCreate(bundle);` we were expecting.

Thus, the breakpoint never hits.

Update these tests to "lookup" the line we want to break on, rather
than hard-coding it.  This way if the file does change slightly, as
long as the expected bit of code is present we can find the line
number we want to break on.

Update the `DotNetNUnitCategories` YAML variable so that the Debugger
test category is no longer excluded on .NET 6.  This ensures that
`ApplicationRunsWithDebuggerAndBreaks` is run as part of the
**MSBuildDeviceIntegration On Device - One .NET** job.