[build] Output templates pack to lowercase path (#7167)
Workload template packs installed into the `template-packs` folder
should have a lowercased package ID.  This was not accounted for when
updating local build behavior in commit c21e78ca.  Fix the path that we
"install" our template pack into for local builds so that the templates
can be resolved on Linux.