[Localization] Add OneLocBuild job (#175)
Context: https://aka.ms/OneLocBuild

A new pipeline has been added in the DevDiv project which will run
the `OneLocBuild@2` task on every commit to main.  This task produces
files needed by the localization team.  For more details on this
process, see the [OneLocBuild documentation][0]

Onboarding for this repo is still pending, and a workflow for the
"handback" of localized files will need to be established in a future
PR once onboarding is complete.

The new pipeline that runs the OneLocBuild task can be found at
<https://devdiv.visualstudio.com/DevDiv/_build?definitionId=16846>.

[0]: https://aka.ms/OneLocBuild