[Localization] Add OneLocBuild job (#7184)
A new job has been added to run the `OneLocBuild@2` task on every
commit to main.  This task produces files needed by the localization
team.  For more details on this process, see the
[OneLocBuild documentation][0].

Onboarding for this repo is still pending, and a workflow for
automatic "handback" of localized files will need to be established
in a future PR once onboarding is complete.

[0]: https://aka.ms/OneLocBuild