[ci] Rename job names
Certain tools such as [Component Governance][0] only use the Pipeline
name -- as defined somewhere in Azure DevOps -- and the `displayName`
property on the [`job` definition][1].  The `displayName` property on
the containing [`stage`][2] is *not* displayed.

This can result in "wonderful" alerts which state that the affected
pipeline is e.g. "Xamarin.Android (macOS)", which isn't particularly
helpful when there are *three* jobs with a `displayName` of `macOS`:

  * `designer_integration_mac`
  * `mac_bcl_tests`
  * `mac_build_create_installers`

Fully understanding these alerts is irksome.

Improve this situation by introducing "redundancy" into the job
`displayName` values, and updating them to follow a pattern:

 1. Host Platform
 2. Category
 3. (Optional) Description

Each part is separated from the others by a ` > `.

Examples:

  * Replace `Mac Build` with `macOS > Build`.
  * For the previous `macOS` displayName:
    * The `designer_integration_mac` job is `macOS > Tests > Designer Integration`
    * The `mac_bcl_tests` job is `macOS > Tests > BCL (Emulator)`
    * The `mac_build_create_installers` job is `macOS > Create Installers`
  * Replace `Build and Smoke Test` with `Windows > Build & Smoke Test`

[0]: https://devdiv.visualstudio.com/DevDiv/_componentGovernance/112013?_a=alerts&typeId=6317076&alerts-view-option=active
[1]: https://docs.microsoft.com/en-us/azure/devops/pipelines/yaml-schema/jobs-job?view=azure-pipelines
[2]: https://docs.microsoft.com/en-us/azure/devops/pipelines/yaml-schema/stages-stage?view=azure-pipelines