[build] Fix workload install for local dependencies
The `InstallManifestAndDependencies` target is used by local builds to
install the runtime dependencies and WorkloadManifest files required for
building net*-android projects.  This broke after a recent .NET SDK
update, as the empty workload that we only use locally is not known to
the .NET SDK.  Fix this by adding our empty workload to the list of
included workloads when building locally.