MAUI app ran with marshal methods for the first time
Unfortunately, 133 out of 182 methods are still registered dynamically,
the reason being that they either return `bool` or take it as one of
their parameters. `bool` isn't a blittable type and thus such methods
cannot be used with `[UnregisteredCallersOnly]`. To move farther, we
need to modify the generator to stop generating native callbacks (just
them) with `bool`.