[Xamarin.Android.Build.Tasks] Pass `hybrid` to AOT, if enabled (#7263)
Fixes: https://github.com/xamarin/xamarin-android/issues/7088

When the AOT hybrid mode is enabled when the `$(AndroidAotMode)`
MSBuild property is set to `Hybrid`, we failed to inform the AOT
compiler about its desired mode by omitting the `hybrid` argument
from the list of options passed to the compiler via `--aot`.

The same would apply to the full AOT mode.

Add `hybrid` and `full` options to the AOT compiler command line, if
enabled via the `$(AndroidAotMode)` MSBuild property.

The mode is set after processing the `$(AndroidAotAdditionalArguments)`
MSBuild property.  The `$(AndroidAotMode)` property should always be
the definitive source of AOT compiler mode, as it specifies the
options directly supported by us.

Note that the AOT compiler doesn't appear to validate options passed
to it too rigorously, so setting multiple modes in some way, may have
weird/invalid effects.  I don't think it's our place to verify the
modes, thus I'm not adding any code to that effect.