[build] disable classic tests on .NET 6 release branches (#7286)
Many of our Windows lanes on `release/6.0.4xx` are failing with:

    Xamarin.Provisioning.ProvisioningException: Failed to provision
    ---> Xamarin.Provisioning.ProvisioningException: Only extension versions greater than that in the instance catalog version (13.0.0.0) can be installed.

This is because the CI machines have VS 17.3 now, and the classic
Xamarin.Android we're building is too old.

For now, let's just disable all these classic tests on this branch. We
can come up with a better idea for this later.