[ci] Split up signing for files to be notarized (#7321)
The signing steps that run on macOS to prepare a PKG for notarization
have been slow and are frequently failing.  Attempt to fix this by
splitting up the files we send to ESRP on macOS into multiple batches to
hopefully put less pressure on the service.

A `signClassicPkgContent` parameter has also been added to allow us to
skip these steps while testing, or when building .NET release branches if
needed.

MicroBuild tooling has also been updated to the latest versions.