[One .NET] fix `dotnet run -c Release`
Context: https://github.com/dotnet/xamarin/issues/26

The following does not work as expected:

    dotnet new android
    dotnet run -c Release

You end up with a `Debug` app in this case.

The way we implemented `dotnet run` was to make `$(RunCommand)` and
`$(RunArguments)` invoke `dotnet build -t:Run`.

We were losing the `$(Configuration)` value in this example.