[Xamarin.Android.Build.Tasks] Feature for excluding Kotlin-related files from apps
Fixes #6920

Add a new ItemGroup `AndroidPackagingOptionsExclude` to the common
targets. This ItemGroup will contain a set of RegEx compatible patterns
which will be used to exlcude certain files from the final apk/aab.

The patterns will need to be Url encoded so that MSBuild does not try
to interpret them.

The main purpose of this is to allow us to exclude certain Kotlin meta-data
files from the final package. These files seem to be excluded by gradle
so we should do the same.