[Xamarin.Android.Build.Tasks] AndroidLinkResources and Styleables (#7306)
Fixes: https://github.com/xamarin/xamarin-android/issues/7194

Context: https://github.com/dotnet/maui/pull/7038

The initial version of `$(AndroidLinkResources)` (9e6ce03c) was too
broad in its removal of Resource classes and fields.  Certain fields
such as `Styleable` arrays were not called using the IL `stsfld`
opcode.  As a result they could not be easily replaced with constant
usage.

However, the linker removed *all* the fields from the `Resource` 
nested types.  This would result in the following error at runtime:

	System.BadImageFormatException: 'Could not resolve field token 0x0400000b'

This was because the `int[]` fields were removed as part of the
linking process. 

Fix this by leaving the `int[]` fields in the `Resource` nested types
instead of removing them.

We can still remove all the other `int` fields.

We now also need to fix up the `Resource` nested type constructors
to replace the `int` field access with the constant values like we do
for the rest of the app.   This was not required previously because
these constructors were removed, but now we have to keep them because
the static array initialization takes place in these constructors.