[One .NET] add `$(ApplicationIcon)` and `$(ApplicationRoundIcon)`
Fixes: https://github.com/xamarin/xamarin-android/issues/7147

This adds `$(ApplicationIcon)` and `$(ApplicationRoundIcon)`, so they
can be set to a value such as `@mipmap/icon` or `@drawable/icon`
instead of putting this value in `AndroidManifest.xml`.

As described in `OneDotNetSingleProject.md`, these properties will
only be used if `$(GenerateApplicationManifest)` is `true`. This is on
by default in .NET 6+, and off by default in Xamarin.Android.

I purposely did not add these new properties to the `android`
template, as I don't know what IDE impact it will have.

However, MAUI projects already have custom MSBuild targets for
`@(MauiIcon)`, so they can start leveraging this property so the value
can be set in one place.

The MAUI template has:

    <MauiIcon Include="Resources\AppIcon\appicon.svg" ForegroundFile="Resources\AppIcon\appiconfg.svg" Color="#512BD4" />

Currently if you rename `@(MauiIcon)`, you have to also update
the `AndroidManifest.xml` and `Info.plist`.

One "gotcha" is that `$(ApplicationIcon)` is an existing property used
for `<Csc/>` for `Win32Icon="$(ApplicationIcon)"`. I cleared the
original property when `$(GenerateApplicationManifest)` is `true` to
avoid C# compiler errors.