[Xamarin.Android.Build.Tasks] XA1030 for AOT + PublishTrimmed=false (#7406)
Fixes: https://github.com/xamarin/xamarin-android/issues/7178

Using this combination on .NET 6+:

	<PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
	  <RunAOTCompilation>true</RunAOTCompilation>
	  <PublishTrimmed>false</PublishTrimmed>
	</PropertyGroup>

Causes apps to crash at runtime with:

	D Mono    : AOT: module Microsoft.Maui.dll.so is unusable (GUID of dependent assembly Xamarin.AndroidX.AppCompat doesn't match (expected '1FC81757-8A70-4D56-93E6-8A635E2C23DE', got 'FD3821D1-CBF4-4956-B930-EA2A5379E18D')).

AOT runs right after the `<ILLink/>` MSBuild task, and so this
combination currently doesn't work when trimming is disabled.

For now, solve this by emitting an `XA1030` error and fail the build.
This way you at least get a reasonable error at build time instead of
at runtime.

I do not know of a scenario when it would be useful to disable
trimming completely during a `Release` + AOT build.  If you are
hitting a trimming/linker issue, you would be better off preserving
a single assembly, type, etc.