[Xamarin.Android.Build.Tasks] Avoid XA5207 for design-time builds
Fixes #7405

Raising a XA5207 Error during a design time build is causing some
problems for the auto sdk installer. The design time build is run for
all frameworks. So you end up in a position when you have the Windows
framework selected in the IDE and the auto sdk installer will try to
install an Android framework. Its very confusing for the user.

So lets not raise this Error during a design time build.