[build] fix 8.0.100-alpha.1 version band
Local builds were failing with:

    make prepare all
    ...
    build-tools/create-packs/Directory.Build.targets(280,5): error MSB3073: The command ""…/xamarin-android/bin/Debug/dotnet/dotnet" workload install android-deps --configfile "…/xamarin-android/NuGet.config" --skip-manifest-update --verbosity diag" exited with code 1.

`android-deps` is our workaround for installing the "abstract" Mono
workload that we need to build against.

Since we are "inventing" this workload, there is not a way for the .NET
workload system to fall back to older version bands.

And so:

    bin/Debug/lib/sdk-manifests/8.0.100/android.deps.workload/WorkloadManifest.json

Doesn't work because the version band is *actually* 8.0.100-alpha.1.

Update our logic in `eng/Versions.props` to support `alpha` version bands.