Remove use of SixLabors.ImageSharp
See also: 24d5280091bdee84f15249c28431d68833c8fbb2

There does not appear to be an easy way to get SixLabors.ImageSharp
working with updated NuGet dependencies, and I wonder… why do we need it?

We use SixLabors.ImageSharp to resize an image for different sizes,
but do we *need* the images to be different sizes?  Or would anything
break if they were all the *same* size?

Try it!  Use the same icon data for all image sizes.