[Build] Log better errors when compiling/linking with binutils
If an error happens when compiling or linking native assembler files
generated by Xamarin.Android, we log an error which doesn't contain much
information:

   Error	XA3006	Could not compile native assembly file: compressed_assemblies.x86.ll

The real cause of the issue is logged as well, however not as an error
but rather a "plain" message, which is not immediately accessible to the
user experiencing the issue.

Keep logging the output messages of the assembler and linker programs as
before, but whenever an error occurs, log their entire output (both
stdout and stderr) as an MSBuild error message.