[Xamarin.Android.Build.Tasks] Useful errors when using binutils (#7566)
Context: b002dc38196a0bedb18c82d228056e1ba108459c
Context: https://github.com/xamarin/xamarin-android/pull/2905#issuecomment-479599104

If an error occurs when compiling or linking native assembler files
generated by .NET Android, we log an error which doesn't contain any
useful information:

	error XA3006: Could not compile native assembly file: compressed_assemblies.x86.ll

The real cause of the issue is logged as well, but not as an error;
it's visible in the Diagnostic log as a "plain" message, which is not
immediately accessible to the user experiencing the issue.

Keep logging the output messages of the assembler and linker programs
as before, but whenever an error occurs, log their entire output
(stdout and stderr) as a single MSBuild message with newlines between
each each line of stdout/stderr.  This will render within Visual Studio
as a multi-line error message ("CreateErrorsItemGroupSingleLine" in
xamarin/xamarin-android#2905).