[ci] Add API-33 to the nightly build (#7552)
Fixes: https://github.com/xamarin/xamarin-android/issues/7490

Add API-33 to the nightly build tests.  Note that the `default` image
is not available so we have to use the `google_apis` image instead. 

Also it seems that under API-33 dotnet is unable to get the hardware
address of the network adapters.  As a result the tests which compare
these were failing.

Ignore the hardware address check on API-33, while continuing to
check other aspects of the network adapters.

API-34+ will *not* ignore the hardware address check.  This will be
reconsidered when API-34 actually exists.